Rework sampling of scalars
As the next step would involve computing the inverse of omega, we needed
to change sampling strategies (the previous representation of omega as a
byte array would not have easily allowed for this).

We now start from BigUints, which allow us to manipulate the values
(such as to find inverses) easily before representing them as byte
arrays. This also allows us to obtain the exact bitlength of these
representations, used later on.

One thing to be careful about: operating over BigUints may not be
constant-time, they may leak information about the generated scalars.